Move scheme comparison plot into plotting_utilities
